Move HashValue and Blob into separate headers, rename `Hmac.h` -> `HmacContext.h`
About to make changes...
